fix accordion in QA
