+++ client/package.json
@@ -24,6 +24,7 @@
     "@babel/preset-typescript": "^7.14.5",
     "@emotion/react": "^11.4.1",
     "@emotion/styled": "^11.3.0",
+    "@mui/icons-material": "^5.0.1",
     "@mui/material": "^5.0.2",
     "@testing-library/jest-dom": "^5.14.1",
     "@types/jest": "^26.0.23",
